more fixes from changes requested
